google: Bump credentials generation
Access to the following were turned on in the Google APIs Console:
 - Calendar API
 - Google Calendar CalDAV API
 - Google Contacts CardDAV API

Of these, only the last two are new. We were already requesting the
scope for Calendar API, but looks like the APIs Console is the way
to go now. Interestingly the APIs Console does not list all the other
services that we are interested in, or it is does but is not obvious
to me.

In any case we need access to their new CalDAV API which works with
OAuth2 because that would let us work with 2-factor authenticated
accounts again.

See: https://bugzilla.gnome.org/show_bug.cgi?id=686804
     https://bugzilla.gnome.org/show_bug.cgi?id=688364