oauth2: Let backends extract the code or token themselves
Certain OAuth2-like, but not exactly OAuth2, services do not follow
the standard mechanisms for extracting the access token or auth code
from the redirect URI. They use some non-standard technique to do so.
Let us try to accommodate them by allowing a provider specific hook
which we will invoke only if it is implemented.

https://bugzilla.gnome.org/show_bug.cgi?id=704564