kerberos: don't crash if keyring credentials disappear
It's possible to make gnome-online-accounts crash by:

1) creating a kerberos account in control-center
2) deleting the stored credentials from gnome keyring using seahorse
3) running kdestroy in the terminal to make the sign in button show up
in the control-center panel
4) clicking sign in

This is because the provider makes the assumption that there will always
be stored credentials when signing in interactively, which is incorrect
in the above scenario (because of step 2).

This commit hardens the get_ticket_sync function against credentials
disappearing.

https://bugzilla.gnome.org/show_bug.cgi?id=711572